lab1/pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.32.0",
    "openai-agents>=0.14.3",
    "pdfplumber>=0.11.9",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
